Refactor helpers and allow passing focus option to open, close and toggle API
